pnpm create monkey + preact
