+++ userscripts/atualizar-saldo/.gitignore
@@ -1,0 +1,1 @@
+dist/
