--- userscripts/atualizar-saldo/metadata.js
@@ -1,17 +1,0 @@
-// ==UserScript==
-// @name        atualizar-saldos
-// @name:pt-BR  Atualizar saldos
-// @namespace   http://nadameu.com.br
-// @match       https://eproc.jfpr.jus.br/eprocV2/controlador.php?acao=processo_precatorio_rpv&*
-// @match       https://eproc.jfrs.jus.br/eprocV2/controlador.php?acao=processo_precatorio_rpv&*
-// @match       https://eproc.jfsc.jus.br/eprocV2/controlador.php?acao=processo_precatorio_rpv&*
-// @match       https://eproc.trf4.jus.br/eproc2trf4/controlador.php?acao=processo_precatorio_rpv&*
-// @match       https://eproc.jfpr.jus.br/eprocV2/controlador.php?acao=processo_selecionar&*
-// @match       https://eproc.jfrs.jus.br/eprocV2/controlador.php?acao=processo_selecionar&*
-// @match       https://eproc.jfsc.jus.br/eprocV2/controlador.php?acao=processo_selecionar&*
-// @match       https://eproc.trf4.jus.br/eproc2trf4/controlador.php?acao=processo_selecionar&*
-// @grant       none
-// @version     2.0.0
-// @author      nadameu
-// @description Atualiza o saldo de contas judiciais
-// ==/UserScript==
