--- userscripts/atualizar-saldo/src/global.d.ts
@@ -1,1 +1,0 @@
-declare const DEVELOPMENT: boolean;
