+++ userscripts/atualizar-saldo/src/paginaProcesso.scss
@@ -1,0 +1,35 @@
+.infra-styles {
+  .gm-atualizar-saldo {
+    button {
+      --regular-bg: hsl(266, 25%, 40%);
+      --highlighted-bg: hsl(266, 40%, 45%);
+
+      margin: 2px 0 8px;
+      border: none;
+      border-radius: 3px;
+      padding: 4px 20px;
+      font-size: 0.86rem;
+      color: white;
+      box-shadow: 0 2px 4px hsla(0, 0%, 0%, 0.5);
+      background: var(--regular-bg);
+      &:hover,
+      &:focus {
+        background: var(--highlighted-bg);
+      }
+      &:active {
+        box-shadow: none;
+        translate: 0 2px;
+        // background: var(--highlighted-bg);
+      }
+    }
+    output {
+      font-size: 1.3em;
+      line-height: 32px;
+      color: hsl(266, 80%, 15%);
+
+      &.erro {
+        color: hsl(0, 85%, 40%);
+      }
+    }
+  }
+}
