+++ userscripts/atualizar-saldo/src/preact.d.ts
@@ -1,0 +1,1 @@
+import JSX = preact.JSX;
