--- userscripts/atualizar-saldo/userscript.json
@@ -1,1 +1,0 @@
-{ "entry": "src/index.ts", "metadata": "metadata.js", "outfile": "../../atualizar-saldo.user.js" }
